Change name of Parse.swift to JSONParser
